refactor: extract core logic of router
